Move hdf/hdfs, another parquet clean
